Added create appointment route
